requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.0
pytubefix>=6.0.0
numpy>=1.26.0

# LangChain ecosystem
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.0.3
langchain-google-genai>=1.0.0
langchain-xai>=0.0.1
langchain-classic>=0.0.1

# Vector store & retrieval
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
flashrank>=0.2.0

# Embeddings
sentence-transformers>=3.0.0

# Whisper transcription (optional, used as caption fallback)
faster-whisper>=1.0.0
static-ffmpeg>=2.5.0

# RAG evaluation
ragas>=0.1.0
datasets>=2.19.0
pandas>=2.0.0
youtube-transcript-api>=0.6.0
